+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @elsium-ai/mcp
 
+## 0.12.1
+
+### Patch Changes
+
+- Updated dependencies [6a0eb78]
+  - @elsium-ai/tools@0.12.1
+  - @elsium-ai/core@0.12.1
+
 ## 0.12.0
 
 ### Patch Changes